tests: prev-next unit testing port
The following ports over the original prev-next tests from the legacy
container into the unit testing root.
